+++ handle_table.h
@@ -17,21 +17,21 @@ typedef struct ClientNode {
     struct ClientNode *next;    // Pointer to the next node in the linked list (Handles collisions w/ chaining)
 } ClientNode;
 
-typedef struct {
+struct HandleTable{
     ClientNode **buckets;    // Array of pointers to linked lists
     int size;                // Size of the table
     int count;               // Number of elements in the table
-} HandleTable; // Hash table structure
+} ; // Hash table structure
 
 /* Function Prototypes */
 unsigned int hash_func(const char *handle, int size);
-HandleTable *create_table(int size);
-void resize_table(HandleTable *table);
-void rehash_table(HandleTable *table, int new_size);
-void addHandleSockPair(HandleTable *table, const char *handle, int socket);
-int getSockNum(HandleTable *table, const char *handle);
-void removeHandleSockPair(HandleTable *table, const char *handle);
-//void print_table(HandleTable *table); // For debugging
+void create_table(int size);
+void resize_table();
+void rehash_table(int new_size);
+void addHandleSockPair(const char *handle, int socket);
+int getSockNum(const char *handle);
+void removeHandleSockPair(const char *handle);
+//void print_table(); // For debugging
 //void free_table(HandleTable *table); // Frees the memory allocated for the table