+++ server.c
@@ -26,9 +26,13 @@
 #include "safeUtil.h"
 #include "pdu_io.h"
 #include "pollLib.h"
+#include "handle_table.h"
 
 /* Definitions */
 #define MAXBUF 1024
+#define MAX_MSG_SIZE 200
+#define MAX_H 101
+#define MAX_NUM_HANDLES 9
 #define DEBUG_FLAG 1
 
 /* Function Prototypes */
@@ -37,7 +41,7 @@ void serverControl(int mainServerSocket, int clientSocket);
 void addNewSocket(int mainServerSocket);
 void processClient(int clientSocket); // renamed from recvFromClient
 void flagCheck(uint8_t *dataBuffer, int clientSocket, int messageLen);
-
+void processMCMsgs(uint8_t *dataBuffer, int clientSocket, int messageLen);
 
 int main(int argc, char *argv[])
 {
@@ -52,6 +56,7 @@ int main(int argc, char *argv[])
 	mainServerSocket = tcpServerSetup(portNumber);
 
 	addToPollSet(mainServerSocket);
+	create_table(INIT_TABLE_SIZE); 
 	while (1) {
 		serverControl(mainServerSocket, clientSocket);	// Handles processing connections
 	}
@@ -80,7 +85,6 @@ int checkArgs(int argc, char *argv[])
 	{
 		portNumber = atoi(argv[1]); 
 	}
-	
 	return portNumber;
 }
 
@@ -88,7 +92,7 @@ int checkArgs(int argc, char *argv[])
 void serverControl(int mainServerSocket, int clientSocket) {
 	// Wait for a socket to be ready
 	int resp_sock = pollCall(-1); // Blocks until a socket is ready
-
+	
 	if (resp_sock < 0) {
 		perror("pollCall");
 		exit(-1);
@@ -126,12 +130,12 @@ void processClient(int clientSocket) // used to be recvFromClient
 		perror("recv call");
 		exit(-1);
 	}
-	printf("Buffer received at server: %s\n", dataBuffer);
+	//printf("Buffer received at server: %s\n", dataBuffer);
 	// Check if there is data to be received
 	if (messageLen > 0)
 	{
 		//printf("Message received on socket: %d, length: %d Data: %s\n", clientSocket, messageLen, dataBuffer);
-		flagCheck(dataBuffer);
+		flagCheck(dataBuffer, clientSocket, messageLen);
 	}
 	// Check if the connection is closed
 	else if (messageLen == 0)
@@ -143,12 +147,21 @@ void processClient(int clientSocket) // used to be recvFromClient
 }
 
 void flagCheck(uint8_t *dataBuffer, int clientSocket, int messageLen) {
-	if (dataBuffer[0] == 5) { // Send message flag
+	if (dataBuffer[0] == 1) { // Add handle flag
+		printf("Initial packet flag received by server.\n");
+		uint8_t handle_len = dataBuffer[1];
+		uint8_t handle[MAX_H] = {0};
+		memcpy(handle, &dataBuffer[2], handle_len);
+		addHandleSockPair((const char *)handle, clientSocket);
+	}
+	else if(dataBuffer[0] == 5) { // Send message flag
 		printf("Send message flag received by server.\n");
 		processMCMsgs(dataBuffer, clientSocket, messageLen);
 	}
 	else if (dataBuffer[0] == 6) { // Multicast flag
 		printf("Multicast flag received by server.\n");
+		processMCMsgs(dataBuffer, clientSocket, messageLen);
+
 	}
 	else if (dataBuffer[0] == 4) { // Broadcast flag
 		printf("Broadcast flag received by server.\n");
@@ -162,46 +175,49 @@ void flagCheck(uint8_t *dataBuffer, int clientSocket, int messageLen) {
 }
 
  /* Processes messages received from server and displays them on client terminal */
-/*void processMCMsgs(uint8_t *dataBuffer, int clientSocket, int messageLen) {
+void processMCMsgs(uint8_t *dataBuffer, int clientSocket, int messageLen) {
 	int packetIndex = 1; // Skip the flag, start at the sender handle/number of handles
+	uint8_t lengthofsenderhandle = dataBuffer[packetIndex++]; // 1 byte
+	packetIndex += lengthofsenderhandle; // Skip the sender handle
+
 	// Parse buffer to check for num handles
 	// Check how many handles are in dataBuffer[packetIndex]
-	uint8_t num_handles = dataBuffer[packetIndex]; // 1 byte
-	packetIndex++; // Increment to the next byte
-	if (num_handles > 1) { // Multicast
-		// Parse buffer for destination handles
-		for (int i = 0; i < num_handles; i++) {
-			uint8_t desthandle_len = dataBuffer[packetIndex]; // 1 byte
-			packetIndex++; // Increment to the next byte
-			
-
-		}
-		// add to handle table 
-		// access handle table to figure what socket to send to
-
-		// Parse buffer for message
-
-		// figure out how to send message to each socket
-	}
-	else if (num_handles == 1) { // Unicast
-		// Parse buffer for destination handles
-		// add to handle table 
-		// access handle table to figure what socket to send to
-
-		// Parse buffer for message
-
-		// figure out how to send message to each socket
-	}
+	uint8_t num_handles = dataBuffer[packetIndex++]; // 1 byte
+	
+	uint8_t lengthofdestinationhandle = 0;
+	uint8_t handleArray[MAX_NUM_HANDLES][MAX_H] = {0}; // Buffer to store destination handles
+	// Parse buffer for destination handles
+	for (int i = 0; i < num_handles && i < MAX_H; i++) {
+		lengthofdestinationhandle = dataBuffer[packetIndex++]; // 1 byte
+		memcpy(handleArray[i], &dataBuffer[packetIndex], lengthofdestinationhandle);
+		packetIndex += lengthofdestinationhandle;
+	};
+	
+	// figure out how to send message to each socket
+	// Iterate through handle array to send message to each socket
+	for (int handle_indx = 0; handle_indx < num_handles; handle_indx++) {
+		
+		uint8_t *current_handle = handleArray[handle_indx];
+		uint8_t current_handle_length = strlen((char *) current_handle);
+		
+		// Get socket number from handle table
+		int destSocket = getSockNum((char *)current_handle);
 	
 
+		if (destSocket == - 1) {
+			// Create "client does not exist" packet
+			uint8_t error_packet[MAX_H + 2] = {0};
+			error_packet[0] = 7; 	// 1 byte Flag
+			error_packet[1] = current_handle_length;	// 1 byte
+			memcpy(&error_packet[1], handleArray[handle_indx], current_handle_length);
+			// Send a "client does not exist" packet to destination handle
+			sendPDU(clientSocket, error_packet, current_handle_length + 2);
+		}
+		else {
+			// Send original packet to destination handle
+			sendPDU(destSocket, dataBuffer, messageLen);
+		}
+		
+	}
 
-	// Parse buffer for destination handles
-	// add to handle table 
-	// access handle table to figure what socket to send to
-
-	// Parse buffer for message
-
-	// figure out how to send message to each socket
-
-
-}*/
+}